+++ hutool-extra/src/main/java/cn/hutool/extra/ssh/Sftp.java
@@ -462,9 +462,13 @@ public class Sftp extends AbstractFtp {
 			fileName = entry.getFilename();
 			if (false == ".".equals(fileName) && false == "..".equals(fileName)) {
 				if (entry.getAttrs().isDir()) {
-					delDir(fileName);
+					// pr#1380Gitee 当目录名包含特殊字符（如 \u000b）时，会导致不断进入同一目录循环
+					// 此处强制使用绝对路径
+					delDir(dirPath + "/" + fileName);
 				} else {
-					delFile(fileName);
+					// pr#1380Gitee 当目录名包含特殊字符（如 \u000b）时，会导致不断进入同一目录循环
+					// 此处强制使用绝对路径
+					delFile(dirPath + "/" + fileName);
 				}
 			}
 		}